refactor: SRP 원칙에 따라 디렉터리 및 메서드 책임 분리
